refactor(airflow): split operators.py into module
