+++ airflow/dags/gtfs_loader/gtfs_schedule_history_load.py
@@ -8,7 +8,7 @@
 
 import pandas as pd
 from calitp import get_fs, get_bucket, get_table
-from operators import _keep_columns
+from utils import _keep_columns
 
 # Note that destination includes date in the folder name, so that we can use
 # a single wildcard to in the external table source url
