+++ airflow/plugins/operators/__init__.py
@@ -1,0 +1,6 @@
+# flake8: noqa
+from .external_table import ExternalTable
+from .once_off_external_task_sensor import OnceOffExternalTaskSensor
+from .pod_operator import PodOperator
+from .python_taskflow_operator import PythonTaskflowOperator
+from .sql_to_warehouse_operator import SqlToWarehouseOperator
