+++ airflow/plugins/utils.py
@@ -1,0 +1,37 @@
+import pandas as pd
+from calitp import read_gcfs, save_to_gcfs
+
+
+def _keep_columns(
+    src_path, dst_path, colnames, itp_id=None, url_number=None, extracted_at=None,
+):
+
+    # read csv using object dtype, so pandas does not coerce data
+    df = pd.read_csv(read_gcfs(src_path), dtype="object")
+
+    if itp_id is not None:
+        df["calitp_itp_id"] = itp_id
+
+    if url_number is not None:
+        df["calitp_url_number"] = url_number
+
+    # get specified columns, inserting NA columns where needed ----
+    df_cols = set(df.columns)
+    cols_present = [x for x in colnames if x in df_cols]
+
+    df_select = df.loc[:, cols_present]
+
+    # fill in missing columns ----
+    print("DataFrame missing columns: ", set(df_select.columns) - set(colnames))
+
+    for ii, colname in enumerate(colnames):
+        if colname not in df_select:
+            df_select.insert(ii, colname, pd.NA)
+
+    if extracted_at is not None:
+        df_select["calitp_extracted_at"] = extracted_at
+
+    # save result ----
+    csv_result = df_select.to_csv(index=False).encode()
+
+    save_to_gcfs(csv_result, dst_path, use_pipe=True)
