+++ airflow/requirements.txt
@@ -2,7 +2,7 @@ gcsfs==0.8.0
 intake==0.6.1
 ipdb==0.13.4
 pandas==1.1.4
-gusty==0.5.1
+gusty==0.6.0
 pandas-gbq==0.14.1
 pybigquery==0.7.0
 google-cloud-bigquery==2.16.1
